requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.8.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
letta
supabase>=2.3.0
httpx>=0.25.2
sse-starlette>=1.8.2
websockets>=12.0
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.3
aiofiles>=23.2.1